+++ CHANGELOG.rst
@@ -2,6 +2,13 @@ Changelog
 =========
 
 
+1.3.0 (2019-04-10)
+------------------
+
+* Allow passing multiple directories on the CLI
+* Allow glob patterns of directories on the CLI
+
+
 1.2.1
 -----